Split commands file into multiple files
